started readme

+++ include/config.hpp
@@ -6,7 +6,7 @@ const unsigned short WINDOW_WIDTH = 800;
 const unsigned short WINDOW_HEIGHT = 600;
 
 /* AGENT */
-const unsigned int AGENT_COUNT = 1024;
+const unsigned short AGENT_COUNT = 1024;
 const float AGENT_MOVE_SPEED = 0.00006f; // As a proportion of the screen per frame
 const float AGENT_TURN_SPEED = 1.0f; // unimplemented
